Improving modularity
