+++ src/util/helper.js
@@ -6,21 +6,6 @@ function getConsole() {
 }
 const console = getConsole();
 
-function cached(fn) {
-  const cache = Object.create(null);
-  return function cachedFn(str) {
-    const hit = cache[str];
-    return hit || (cache[str] = fn(str));
-  };
-}
-
-const regex = /-(\w)/g;
-const camelize = cached(str =>
-  str.replace(regex, (_, c) => (c ? c.toUpperCase() : ""))
-);
-
-const capitalize = str => str.replace(/^\w/, c => c.toUpperCase());
-
 function removeNode(node) {
   if (node.parentElement !== null) {
     node.parentElement.removeChild(node);
@@ -35,4 +20,4 @@ function insertNodeAt(fatherNode, node, position) {
   fatherNode.insertBefore(node, refNode);
 }
 
-export { insertNodeAt, camelize, capitalize, console, removeNode };
+export { insertNodeAt, console, removeNode };
