+++ src/util/string.js
@@ -1,0 +1,17 @@
+function cached(fn) {
+  const cache = Object.create(null);
+  return function cachedFn(str) {
+    const hit = cache[str];
+    return hit || (cache[str] = fn(str));
+  };
+}
+
+const regex = /-(\w)/g;
+const camelize = cached(str =>
+  str.replace(regex, (_, c) => (c ? c.toUpperCase() : ""))
+);
+
+const capitalize = str => str.replace(/^\w/, c => c.toUpperCase());
+
+
+export { camelize, capitalize };
