+++ tests/unit/util/helper.spec.js
@@ -1,21 +1,4 @@
-import { camelize, console } from "@/util/helper";
-
-describe("camelize", () => {
-  test.each([
-    ["MyProp", "MyProp"],
-    ["MyProp", "MyProp"],
-    ["kebab-case", "kebabCase"],
-    ["multi-hyphen-string", "multiHyphenString"],
-    ["drag-class", "dragClass"],
-    ["test-", "test-"]
-  ])(
-    "transform %s into %s",
-    (value, expected) =>{
-      const actual = camelize(value);
-      expect(actual).toEqual(expected);
-    }
-  )
-});
+import { console } from "@/util/helper";
 
 describe("console", () => {
   test.each([
